+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.1.98
 
+- Crunchyroll: fixed release calendar bug: saves premiere shows in local list.
 - Hotstar: a lot of bug fixes by @Sathvik6323
 
 ## 1.1.97